requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
osmnx==1.3.0
pyrosm
shapely==2.0.1
polyline
scikit-learn